requirements.txt
openai
python-dotenv
faiss-cpu
sentence-transformers
PyMuPDF
numpy
